==> analysis/part2_forecASD.Rmd <==
---
title: "network model"
output:
  workflowr::wflow_html:
    code_folding: hide
    toc: true
date: "2024-11-05"
---


```{r, message=F, warning=F}
# code from https://github.com/LeoBman/forecASD/blob/master/02_network_models.R
# # # # # # # # # # # # 
#### Load Packages ####
# # # # # # # # # # # #
rm(list=ls())
set.seed(123)
require(randomForest)

```



```{r, message=F, warning=F, eval=T}
path="C:\\han\\Projects\\IntegratePPI\\"
# # # # # # # # # # # #
#### BrainSpan RF #####
# # # # # # # # # # # #
load(paste(path, "01_BrainSpan_matrix.Rdata", sep=""))
load(paste(path, "01_training_labels.Rdata", sep=""))

bs.x = bs[ rownames(bs) %in% c(pos,neg) , ]
bs.y = as.factor(rownames(bs.x) %in% pos)

## train random forest
set.seed(5136)
bs.rf = randomForest(y=bs.y,
                     x=bs.x,
                     ntree=1000,
                     importance=T,
                     strata=bs.y,
                    # sampsize=c(76,76),
                     sampsize=c(min(table(bs.y)),min(table(bs.y))),   # although there are 674 positive examples, i.e. risk genes, the samplesize must not exceed minimum of each class min(table(bs.y))
                     do.trace=10,
                     proximity=T)

## get predictions on all remaining genes
bs.z =  na.roughfix(bs[ !rownames(bs) %in% rownames(bs.x) , ])

bs.prd = predict(bs.rf, bs.z, type="prob")
bs.prd = rbind(bs.rf$vote, bs.prd)

#save(bs.rf, bs.prd, file="C:\\han\\Projects\\IntegratePPI\\02_brainspan_RF.Rdata")

```





```{r, message=F, warning=F, eval=F}
# # # # # # # # # # # # 
#### the STRING RF ####
# # # # # # # # # # # # 
################################## DO NOT RUN-- too long time ###########
load(paste(path, "tmp01_STRING_graph.Rdata", sep=""))

string.x = string.path[rownames(string.path) %in% c(pos,neg), ]
string.y = as.factor( rownames(string.x) %in% pos )

set.seed(2176)
string.rf = randomForest(
  y = string.y,
  x = string.x,
  importance = T,
  strata = string.y,
  #sampsize = c(77,77),
  sampsize = c(min(table(string.y)),min(table(string.y))),  # the samplesize must not exceed minimum of each class min(table(bs.y)) 
  ntree=500, 
  do.trace=10,
  proximity=T)

## stratify by predicted class lables
strata = string.rf$y:string.rf$pred
## sample size is smallest class in strata
sampsize = rep(min(table(strata)),4)

## fit a model while omitting variables which are not used in any tree
set.seed(679)
while( sum( varUsed(string.rf) < 1 ) > 0 ){
 string.rf = randomForest(
   y = string.y,
   x = string.x[, rownames(string.rf$importance)[varUsed(string.rf) > 0] ],
   importance = T,
   strata = strata,
   sampsize = sampsize,
   ntree = 500,
   do.trace = 100,
   proximity = T)
 
 print(paste(sum(varUsed(string.rf) == 0), "out of", nrow(string.rf$importance), "variables unused"))
}


string.prd = predict(string.rf, string.path[!rownames(string.path) %in% rownames(string.x), ], type="prob")
string.prd = rbind(string.prd, string.rf$vote)
string.prd = string.prd[ order(string.prd[,2], decreasing=T), ]

#save(string.rf,string.prd,file=paste(path, "02_STRING_rf.Rdata", sep=""))

```









```{r, message=F, warning=F, eval=F}
# # # # # # # # # # # # # #
#### integrate scores #####
# # # # # # # # # # # # # #
## the STRING RF score, the BrainSpan RF score, and the TADA score


load(paste(path, "02_STRING_rf.Rdata", sep=""))
load(paste(path, "01_id_conversion.Rdata", sep=""))

nn = rownames(string.prd)[rownames(string.prd)%in%rownames(bs.prd)]
symbol = eg.map[names(e2e)[match(nn,e2e)]]



x.pred = data.frame(
  string = string.prd[nn, "TRUE"],
  brainspan = bs.prd[nn, "TRUE"]
  )
x.pred = na.roughfix(x.pred)

x.train = x.pred[rownames(x.pred) %in% c(pos,neg), ]
y.train = as.factor(rownames(x.train) %in% pos)


set.seed(5393)
network.rf = 
  randomForest(
    y = y.train,
    x = x.train[,c("string", "brainspan")],
    importance = T,
    strata = y.train,
    sampsize = c(min(table(y.train)),min(table(y.train))),
    do.trace = 50,
    ntree = 1000)

network.prd = predict(network.rf,
                          x.pred[!rownames(x.pred) %in% rownames(network.rf$votes), c("string", "brainspan")],
                          type=  "prob")
network.prd = rbind(network.prd, network.rf$votes)

#save(network.rf, network.prd, file=paste(path, "02_network_rf.Rdata"))
```
