==> analysis/part3_forecASD.Rmd <==
---
title: "ensemble model"
output:
  workflowr::wflow_html:
    code_folding: hide
    toc: true
date: "2024-11-05"
---


```{r, message=F, warning=F, eval=T}
# # # # # # # # # # # # #
#### Ensembel Model #####
# # # # # # # # # # # # #
rm(list=ls())
set.seed(123)
library(randomForest)

path="C:\\han\\Projects\\IntegratePPI\\"

load(paste(path, "01_training_labels.Rdata", sep=""))
load(paste(path, "02_STRING_rf.Rdata", sep=""))
load(paste(path, "02_brainspan_RF.Rdata", sep=""))
load(paste(path, " 02_network_rf.Rdata", sep=""))

```






```{r, message=F, warning=F, eval=T}

# # # # # # # # # # # # 
#### Load in data #####
# # # # # # # # # # # # 

#meta = read.csv("./ext_data/composite_table.csv", stringsAsFactors = F, row.names = 1)
meta = read.csv("C:\\han\\Dataset\\20240610_PMID_33244169\\composite_table.csv",stringsAsFactors = F, row.names = 1)

############### unify row names among 3 data sets 
common_rownames=intersect(rownames(bs.prd), intersect(rownames(string.prd), rownames(meta)))
string.prd=string.prd[rownames(string.prd) %in% common_rownames, ]
bs.prd=bs.prd[rownames(bs.prd) %in% common_rownames, ]
meta=meta[rownames(meta) %in% common_rownames, ]


if (!all(rownames(meta) %in% rownames(string.prd))) {
  stop("Mismatch in rownames between meta and string.prd")
}

#string.prd <- string.prd[rownames(meta), ]



## combine other predictors with network scores
meta = cbind(
  data.frame(
    STRING_score = string.prd[common_rownames , "TRUE"],
    BrainSpan_score = bs.prd[common_rownames, "TRUE"]
  ),
  meta[rownames(meta), ]
)

```





```{r, message=F, warning=F}

# # # # # # # # # # # # 
#### train forest #####
# # # # # # # # # # # # 

meta.train = na.roughfix(
  meta[meta$ensembl_string %in% c(pos,neg), -(3:9)]   ## remove gene identifiers, etc. use all other information to train the model 
) 
y = as.factor(rownames(meta.train) %in% pos)

set.seed(43775)
rf = randomForest(
  y = y,
  x = meta.train, 
  importance = T,
  do.trace = 10,
  strata = y,
  sampsize = c(min(table(y)),min(table(y)))
  )

meta.test = na.roughfix(
  meta[!(meta$ensembl_string %in% c(pos,neg)), -(3:9)]   ## remove gene identifiers, etc.
) 

meta.prd <- predict(rf, 
        meta.test, 
        type = "prob")

meta.score <- rbind(rf$votes, meta.prd)

final.data <- cbind(
  data.frame(
    forecASD = meta.score[rownames(meta),"TRUE"],
    'STRING+BrainSpan_RF' = network.prd[rownames(meta),"TRUE"]
  ), 
  meta
)

#write.csv(final.data, 
#          file = paste(path, "forecASD_table.csv", sep=""),
#          quote = F, 
#          row.names = F)

```
