==> analysis/part3_hmrfm.Rmd <==
---
title: "HMRFM"
output:
  workflowr::wflow_html:
    code_folding: hide
    toc: true
date: "2024-10-02"
---



## selecting incidence matrix 


```{r, message=F, warning=F, eval=T}
rm(list=ls())
set.seed(123)
require(statnet) # load the package statnet to visualize networks in R
library(ggplot2)
library(dplyr)
# curate the adj_pfc35 by incorporating PPI information 
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\adj_pfc35_for_tada2022.RData")  # load incidence matrix 
load("C:\\han\\Projects\\IntegratePPI\\string.keep_gene_level.RData")  # load pairwise protein interaction information  


### build incidence matrix, (i,j)=1 if there is a link between genes, i & j, and 0 otherwise

#string.keep_gene_level_new=string.keep_gene_level_new %>% filter(combined_score>=800) # apply score threshold 

#string_incidence=matrix(0, nrow=nrow(adj_pfc35), ncol=nrow(adj_pfc35))
#for (i in 1:(nrow(adj_pfc35)-1))
#  for (j in (i+1):nrow(adj_pfc35))
#    if (string.keep_gene_level_new %>% filter(protein1_gene==colnames(dat_pfc35)[i], protein2_gene==colnames(dat_pfc35)[j]) %>% nrow()>0)
#      { 
#      cat(i, "th gene is running", "\n")
#      string_incidence[i,j]=1
#      string_incidence[j,i]=1
#     }
  
#save(string_incidence, file ="C:\\han\\Projects\\IntegratePPI\\string_incidence_with_scorecutoff800.RData")  
load("C:\\han\\Projects\\IntegratePPI\\string_incidence_with_scorecutoff800.RData")
        
###### combine estimated incidence matrix and string PPI information, a link between two genes exists if it appears in estimated incidence matrix or string PPI  
adj_string_combine_incidence=adj_pfc35+string_incidence
for (i in 1:(nrow(adj_string_combine_incidence)-1))
  for (j in (i+1):ncol(adj_string_combine_incidence))
    if (adj_string_combine_incidence[i,j]>1)
    {
      adj_string_combine_incidence[i,j]=1
      adj_string_combine_incidence[j,i]=1
    }


net=network(as.matrix(adj_string_combine_incidence), directed=FALSE) # https://web.stanford.edu/class/bios221/labs/networks/lab_7_networks.html
set.seed(1)          # make the plot look the same as mine
par(mar=rep(0,4))

plot(net, vertex.cex=0.15, rescale=F) # vextex.cex control the vertex size

#dev.off()


```


* adj_pfc35 has 1136 links 

* with score cutoff of 800 to STRING protein interactions, string database has 672 links and combined adj_pfc35 and string has 1804 links 





## run HMRF to get risk genes 




```{r, message=F, warning=F}
fdr_cutoff <- 0.05
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\adj_pfc35_for_tada2022.RData")
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\screen_res_for_tada2022.RData")
#load tada dataset
Jack_fu_data=read.csv("C:\\han\\Dataset\\20231221_PMID_35982160\\Fu 2022 Suppl-41588_2022_1104_MOESM3_ESM.csv")
asd_185_genes=Jack_fu_data$gene[which(Jack_fu_data$FDR_TADA_ASD<0.05)]  # 185 genes are selected at FDR<0.05 https://www.nature.com/articles/s41588-022-01104-0
sum(asd_185_genes %in% tada_2022$Gene)    # among 185 asd genes, 172 appear in brainspan 
```


```{r, message=F, warning=F, eval=F}
# run the HMRF
fdr_cutoff <- 0.05
pthres <- 0.05
source("C:\\han\\Projects\\Multi_Omics\\data_in_MOBS\\hmrf.r")
source("C:\\han\\Projects\\Multi_Omics\\data_in_MOBS\\misc.r")
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\adj_pfc35_for_tada2022.RData")
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\screen_res_for_tada2022.RData")
seedindex <- rep(0, ncol(adj_pfc35))
#seedindex[which(tada$dn.LoF >= 3)] <- 1
seedindex[which(tada_2022$qvalue<1e-06)] <- 1

#tada_2022$pval.TADA[tada_2022$pval.TADA==0]=1e-10
#hmrf_pfc35 <- covarianceSelection::hmrf(tada$pval.TADA, adj_pfc35, seedindex, pthres = pthres)
hmrf_pfc35 <- hmrf(tada_2022$pval.TADA, adj_pfc35, seedindex, pthres = pthres)
#hmrf_pfc35_string <- hmrf(tada_2022$pval.TADA, adj_pfc35, seedindex, pthres = pthres)
#report_pfc35 <- covarianceSelection::report_results(tada$Gene, 1-hmrf_pfc35$post, tada$pval.TADA, hmrf_pfc35$Iupdate)
report_pfc35 <- report_results(tada_2022$Gene, 1-hmrf_pfc35$post, tada_2022$pval.TADA, hmrf_pfc35$Iupdate)
genes_pfc35 <- sort(as.character(report_pfc35$Gene[which(report_pfc35$FDR <= fdr_cutoff)]))
length(genes_pfc35)
#save(genes_pfc35, adj_pfc35, file="C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\result_FDR0.05_adj_pfc35.RData")
length(intersect(asd_185_genes, genes_pfc35))
missed_gene=setdiff(intersect(asd_185_genes,tada_2022$Gene), intersect(asd_185_genes, genes_pfc35))
tada_2022 %>% filter(Gene%in%missed_gene)
```


```{r, message=F, warning=F}
# seedindex significantly influences the estimated risk genes by setting the threshold lambda seedindex[which(tada_2022$qvalue<lambda)] <- 1, so be cautions when setting the seeds!!!!!!! 

ggplot(data.frame(lambda=c(1e-3, 1e-4,1e-05, 1e-6), num_risk_genes=c(572, 471, 354, 237)), aes(x=lambda, y=num_risk_genes, group=1))+
geom_point()+
  geom_line(color="red")


```

* initial seed threshold $\lambda$ signficantly influences the number of risk genes estimated 



```{r, message=F, warning=F, eval=F}
# run the HMRF
fdr_cutoff <- 0.05
pthres <- 0.05
source("C:\\han\\Projects\\Multi_Omics\\data_in_MOBS\\hmrf.r")
source("C:\\han\\Projects\\Multi_Omics\\data_in_MOBS\\misc.r")
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\adj_pfc35_for_tada2022.RData")
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\screen_res_for_tada2022.RData")
seedindex <- rep(0, ncol(adj_pfc35))
#seedindex[which(tada$dn.LoF >= 3)] <- 1
seedindex[which(tada_2022$qvalue<1e-06)] <- 1

#hmrf_pfc35 <- covarianceSelection::hmrf(tada$pval.TADA, adj_pfc35, seedindex, pthres = pthres)
hmrf_pfc35_string <- hmrf(tada_2022$pval.TADA, string_incidence, seedindex, pthres = pthres)
#report_pfc35 <- covarianceSelection::report_results(tada$Gene, 1-hmrf_pfc35$post, tada$pval.TADA, hmrf_pfc35$Iupdate)
report_pfc35_string <- report_results(tada_2022$Gene, 1-hmrf_pfc35_string$post, tada_2022$pval.TADA, hmrf_pfc35_string$Iupdate)
genes_pfc35_string <- sort(as.character(report_pfc35_string$Gene[which(report_pfc35_string$FDR <= fdr_cutoff)]))

length(genes_pfc35_string)
length(intersect(asd_185_genes, genes_pfc35_string))
#save(genes_pfc35_string, string_incidence, file="C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\result_FDR0.05_string_incidence.RData")
```



```{r, message=F, warning=F, eval=F}
# run the HMRF
fdr_cutoff <- 0.05
pthres <- 0.05
source("C:\\han\\Projects\\Multi_Omics\\data_in_MOBS\\hmrf.r")
source("C:\\han\\Projects\\Multi_Omics\\data_in_MOBS\\misc.r")
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\adj_pfc35_for_tada2022.RData")
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\screen_res_for_tada2022.RData")
seedindex <- rep(0, ncol(adj_pfc35))
#seedindex[which(tada$dn.LoF >= 3)] <- 1
seedindex[which(tada_2022$qvalue<1e-06)] <- 1

#hmrf_pfc35 <- covarianceSelection::hmrf(tada$pval.TADA, adj_pfc35, seedindex, pthres = pthres)
hmrf_pfc35_string_combine_adj <- hmrf(tada_2022$pval.TADA, adj_string_combine_incidence, seedindex, pthres = pthres)
#report_pfc35 <- covarianceSelection::report_results(tada$Gene, 1-hmrf_pfc35$post, tada$pval.TADA, hmrf_pfc35$Iupdate)
report_pfc35_string_combine_adj <- report_results(tada_2022$Gene, 1-hmrf_pfc35_string_combine_adj$post, tada_2022$pval.TADA, hmrf_pfc35_string_combine_adj$Iupdate)
genes_pfc35_string_combine_adj <- sort(as.character(report_pfc35_string_combine_adj$Gene[which(report_pfc35_string_combine_adj$FDR <= fdr_cutoff)]))

length(genes_pfc35_string_combine_adj)
length(intersect(asd_185_genes, genes_pfc35_string))
#save(genes_pfc35_string_combine_adj, adj_string_combine_incidence, file="C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\result_FDR0.05_adj_string_combine_incidence.RData")
```




```{r, message=F, warning=F}
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\result_FDR0.05_adj_pfc35.RData")
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\result_FDR0.05_string_incidence.RData")
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\result_FDR0.05_adj_string_combine_incidence.RData")

length(intersect(asd_185_genes, genes_pfc35))
length(setdiff(intersect(asd_185_genes, tada_2022$Gene),genes_pfc35 ))   
missed_genes=setdiff(intersect(asd_185_genes, tada_2022$Gene),genes_pfc35 )
tada_2022 %>% filter(Gene %in% missed_genes)

length(intersect(asd_185_genes, genes_pfc35_string))
length(setdiff(intersect(asd_185_genes, tada_2022$Gene),genes_pfc35_string))  

length(intersect(genes_pfc35, genes_pfc35_string_combine_adj))
length(setdiff(intersect(asd_185_genes, tada_2022$Gene),genes_pfc35_string_combine_adj)) 

```
