==> analysis/part4.enrichment.Rmd <==
---
title: "enrichment analysis"
output:
  workflowr::wflow_html:
    code_folding: hide
    toc: true
date: "2024-10-08"
---


```{r, echo=F, message=F, warning=F}
rm(list = ls())
set.seed(123)
library(VennDiagram)
library(ggplot2)
library(tidyverse)
```



```{r, message=F, warning=F}
#load tada dataset
Jack_fu_data=read.csv("C:\\han\\Dataset\\20231221_PMID_35982160\\Fu 2022 Suppl-41588_2022_1104_MOESM3_ESM.csv")
asd_185_genes=Jack_fu_data$gene[which(Jack_fu_data$FDR_TADA_ASD<0.05)]  # 185 genes are selected at FDR<0.05 https://www.nature.com/articles/s41588-022-01104-0

load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\result_FDR0.05_adj_pfc35.RData")
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\result_FDR0.05_string_incidence.RData")
load("C:\\han\\Projects\\Multi_Omics\\JackFu_2022\\result_FDR0.05_adj_string_combine_incidence.RData")

## SFARI gene set 
SFARI_data=read.csv("C:\\han\\Dataset\\SFARI_base\\SFARI-Gene_genes_08-19-2024release_09-16-2024export.csv", header=T)

sfari_gene=SFARI_data %>% filter(gene.score==1 | gene.score==2) %>% select(gene.symbol) %>% pull() # extract sfari genes in category 1 and 2 
```


## estimated risk genes 

```{r, message=F, warning=F}
ggplot(data.frame(gene_set=c("Autism genes","DAWN genes" , "STRING genes", "DAWN combined STRING genes", "SFARI genes"), num_genes=c(length(asd_185_genes), length(genes_pfc35), length(genes_pfc35_string), length(genes_pfc35_string_combine_adj), length(sfari_gene))), aes(x=gene_set, y=num_genes, fill=gene_set))+
  geom_bar(position = "dodge",stat="identity")+
   xlab("")+ylab("No. genes")+
 #  ylim(c(0,1))+
   theme(axis.text.x = element_text(angle =10, vjust = 0.5, size=10))+
   ggtitle("")+
   geom_text(aes(label=num_genes), position=position_dodge(width=0.9), vjust=-0.25)+  # add numbers over bars 
  #geom_signif(annotations ="p=0.074", y_position = 0.52  ,xmin=1, xmax=2, size=1, vjust=-0.5, textsize=3.5, color="black")+ # textsize is for text over the bar; size is for width of lines 
  # geom_signif(annotations ="p<0.001", y_position = 0.59  ,xmin=1, xmax=3, size=1, vjust=-0.5, textsize=3.5, color="red")+
  #geom_signif(annotations ="p<0.001", y_position = 0.86  ,xmin=1, xmax=4, size=1, vjust=-0.5, textsize=3.5, color="red")+
   theme(plot.title = element_text(hjust = 0.5, size=15))+  #center the title
  theme(legend.title=element_blank())+
  theme(legend.position = "none")
```





##  Venn diagrams of different gene sets 

```{r, message=F, warning=F}
x=list(
  Autism_185_genes=asd_185_genes,
  genes_pfc35=genes_pfc35,
  genes_pfc35_string=genes_pfc35_string,
  genes_pfc35_string_combine_adj=genes_pfc35_string_combine_adj

)

display_venn <- function(x, ...){
  library(VennDiagram)
  grid.newpage()
  venn_object <- venn.diagram(x, filename = NULL, ...)
  grid.draw(venn_object)
}

#png("C:\\Users\\hans\\OneDrive - Marquette University\\SCZ-DAWN\\Figures\\Fig_overlapping_genes.png")

display_venn(
        x,
    #    category.names = c("MIRAGE FDR genes" , "Autism genes" , "DD/ID genes", "SCZ GWAS genes"),
    category.names = c("Autism genes" , "DAWN genes" , "STRING genes", "DAWN combined STRING genes"),
        # Circles
        lwd = 2,
        lty = 'blank',
     #   fill = c("#999999", "#E69F00", "#56B4E9", "#009E73"),
    fill = c("red", "blue", "green", "yellow"),
        # Numbers
        cex = 1.5,
        fontface = "italic",
        # Set names
        cat.cex = 1.2,
        cat.fontface = "bold",
        cat.default.pos = "outer",
       # cat.dist = c(0.055, 0.055, 0.1, 0.1, 0.3 )
         cat.dist = c(0.055, 0.055, 0.1, 0.1 )
)
```




```{r, message=F, warning=F}
x=list(
  sfari_genes=sfari_gene,
  genes_pfc35=genes_pfc35,
  genes_pfc35_string=genes_pfc35_string,
  genes_pfc35_string_combine_adj=genes_pfc35_string_combine_adj

)

display_venn <- function(x, ...){
  library(VennDiagram)
  grid.newpage()
  venn_object <- venn.diagram(x, filename = NULL, ...)
  grid.draw(venn_object)
}

#png("C:\\Users\\hans\\OneDrive - Marquette University\\SCZ-DAWN\\Figures\\Fig_overlapping_genes.png")

display_venn(
        x,
    #    category.names = c("MIRAGE FDR genes" , "Autism genes" , "DD/ID genes", "SCZ GWAS genes"),
    category.names = c("SFARI genes" , "DAWN genes" , "STRING genes", "DAWN combined STRING genes"),
        # Circles
        lwd = 2,
        lty = 'blank',
     #   fill = c("#999999", "#E69F00", "#56B4E9", "#009E73"),
    fill = c("red", "blue", "green", "yellow"),
        # Numbers
        cex = 1.5,
        fontface = "italic",
        # Set names
        cat.cex = 1.2,
        cat.fontface = "bold",
        cat.default.pos = "outer",
       # cat.dist = c(0.055, 0.055, 0.1, 0.1, 0.3 )
         cat.dist = c(0.055, 0.055, 0.1, 0.1 )
)
```


## enrichment with 185 autism genes


* among 185 autism genes reported in Jack 2022 paper, 169 appear in Brainspan data 

* among 169 autism genes, 164 overlap with 225 DAWN genes , 52 overlap with 144 STRING genes,  80 overlap with 243 DAWN combined STRING genes    

* p values 

```{r, message=F, warning=F}
fisher.test(matrix(c(3500, 172, 150, 60), nrow=2))  # p value of autism genes overlapping with STRING 
fisher.test(matrix(c(3500, 172, 237, 172), nrow=2))  # p value of autism genes overlapping with DAWN
fisher.test(matrix(c(3500, 172, 250, 87), nrow=2))  # p value of autism genes overlapping with DAWN combined STRING
```

```{r, message=F, warning=F}
ggplot(data.frame(gene_set=c("DAWN genes" , "STRING genes", "DAWN combined STRING genes"), OR=c(14.75, 8.13, 7.07)), aes(x=gene_set, y=OR, fill=gene_set))+
  geom_bar(position = "dodge",stat="identity")+
   xlab("")+ylab("Odds ratio")+
 #  ylim(c(0,1))+
   theme(axis.text.x = element_text(angle =0, vjust = 0.5, size=12))+
   ggtitle("")+
   geom_text(aes(label=OR), position=position_dodge(width=0.9), vjust=-0.25)+  # add numbers over bars 
  #geom_signif(annotations ="p=0.074", y_position = 0.52  ,xmin=1, xmax=2, size=1, vjust=-0.5, textsize=3.5, color="black")+ # textsize is for text over the bar; size is for width of lines 
  # geom_signif(annotations ="p<0.001", y_position = 0.59  ,xmin=1, xmax=3, size=1, vjust=-0.5, textsize=3.5, color="red")+
  #geom_signif(annotations ="p<0.001", y_position = 0.86  ,xmin=1, xmax=4, size=1, vjust=-0.5, textsize=3.5, color="red")+
   theme(plot.title = element_text(hjust = 0.5, size=15))+  #center the title
  theme(legend.title=element_blank())+
  theme(legend.position = "none")
```



## evidence of new autism risk genes
